requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
dj-database-url==2.1.0
Django==4.2.7
django-admin-interface==0.26.1
django-bootstrap-v5==1.0.11
django-colorfield==0.10.1
django-easy-select2==1.5.8
django-extensions==3.2.3
django-geojson==4.0.0
django-leaflet==0.29.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-slugify==8.0.1
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
